Extract First Status Fns To DB Layer
