+++ controllers/embeddings_status.js
@@ -1,5 +1,5 @@
-const {query, db} = require("@arangodb");
-const {embeddingsStatusCollectionName} = require("../model/embeddings_status");
+const {getStatusByKey} = require("../db/embeddings_status");
+const {getStatusesByCollectionAndEmbName} = require("../db/embeddings_status");
 const {retrieveModel} = require("../services/model_metadata_service");
 const {getEmbeddingsFieldName} = require("../services/emb_collections_service");
 const {sendInvalidInputMessage} = require("../utils/invalid_input");
@@ -13,13 +13,11 @@ function embeddingsStatusesForModel(req, res) {
             `Invalid model: ${modelName} of type ${modelType}`);
     }
 
-    const col = db._collection(embeddingsStatusCollectionName);
-    const statuses = query`
-    FOR d in ${col}
-        FILTER d.collection == ${collectionName}
-        AND d.emb_field_name == ${getEmbeddingsFieldName(fieldName, modelMetadata)}
-        RETURN d
-    `.toArray();
+    const statuses = getStatusesByCollectionAndEmbName(
+        collectionName,
+        getEmbeddingsFieldName(fieldName, modelMetadata)
+    );
+
     if (statuses.length == 0) {
         res.throw(404, "Status not found");
     }
@@ -29,16 +27,11 @@ function embeddingsStatusesForModel(req, res) {
 function embeddingsStatusById(req, res) {
     const {statusId} = req.pathParams;
 
-    const col = db._collection(embeddingsStatusCollectionName);
-    const statuses = query`
-    FOR d in ${col}
-        FILTER d._key == ${statusId}
-        RETURN d
-    `.toArray();
-    if (statuses.length == 0) {
+    const status = getStatusByKey(statusId);
+    if (status == null) {
         res.throw(404, "Status not found");
     }
-    res.json(statuses[0]);
+    res.json(status);
 }
 
 exports.embeddingsStatusesForModel = embeddingsStatusesForModel;
